use uint16_t for fk_item_t.type, uint32_t for fk_item_t.ref

+++ fk_item.h
@@ -1,6 +1,8 @@
 #ifndef _FK_ITEM_H_
 #define _FK_ITEM_H_
 
+#include <stdint.h>
+
 #define FK_ITEM_NIL 	0
 #define FK_ITEM_STR  	1
 #define FK_ITEM_LIST 	2
@@ -8,13 +10,13 @@
 #define FK_ITEM_SKLIST	4
 
 typedef struct {
-    int type;
-    unsigned ref;/* reference count */
+    uint16_t type;
+    uint32_t ref;/* reference count */
     void *entity;/* point to the real data of the itm */
 } fk_item_t;
 
 void fk_item_init(void);
-fk_item_t *fk_item_create(int type, void *entity);
+fk_item_t *fk_item_create(uint16_t type, void *entity);
 void fk_item_dec_ref(fk_item_t *itm);
 void fk_item_inc_ref(fk_item_t *itm);
